Split the dynamic-DNS watcher out of this branch
It lands separately on top.  public_ip keeps the storage half (the DB is authoritative, config
seeds it) since the zone rendering needs it; IP detection goes with the watcher that uses it.

Also restores the start_caddy/elif pairing.  The watcher block had been inserted between
'if config.start_caddy:' and its 'elif needs_caddy_for_tls: raise', so the elif bound to
'if config.dynamic_dns_enabled:' instead — a TLS space with start_caddy on and dynamic DNS off
(the default) would have failed to boot with 'start_caddy is False'.
