--- compute_space/src/compute_space/core/dns/dynamic_dns.py
@@ -1,94 +1,0 @@
-"""Dynamic DNS: keep the instance's A records pointing at wherever it actually is.
-
-Opt-in (``dynamic_dns_enabled``) — on a fixed address the polling is pure cost, but on a
-connection that gets renumbered it is the only thing that brings the space back.
-
-When the router serves DNS the address records are derived from the stored IP at render time, so
-updating them is a re-render.  With an external provider there is nothing to re-render, so the
-records are written through the ``dns`` service instead.
-"""
-
-from __future__ import annotations
-
-import asyncio
-import sqlite3
-from collections.abc import Callable
-from contextlib import closing
-
-from compute_space.config import Config
-from compute_space.core.dns.client import DnsClient
-from compute_space.core.dns.client import router_managed_domains
-from compute_space.core.dns.client import uses_local_dns
-from compute_space.core.dns.coredns_provider.coredns import reload_coredns_for_domains
-from compute_space.core.dns.public_ip import detect_public_ip
-from compute_space.core.dns.public_ip import effective_public_ip
-from compute_space.core.dns.public_ip import store_public_ip
-from compute_space.core.dns.service_api import RecordType
-from compute_space.core.logging import logger
-
-_DEFAULT_INTERVAL_SECONDS = 300.0
-
-# Much shorter than the zone default: pointless to poll every few minutes if resolvers cache the
-# old address for an hour.
-_DYNAMIC_TTL_SECONDS = 60
-
-
-async def check_once(config: Config, db: sqlite3.Connection) -> str | None:
-    """Detect the public IP and, if it moved, store it and re-point the address records.
-
-    Returns the new address when something changed.  A detection failure is not a change: stale
-    records beat pointing the space at nothing.
-    """
-    detected = await detect_public_ip()
-    if detected is None or detected == effective_public_ip(config, db):
-        return None
-
-    logger.info(f"Public IP changed to {detected}")
-    store_public_ip(db, detected)
-
-    if uses_local_dns(db):
-        # Re-rendering *is* the update: the address records come from the stored IP.  Writing them
-        # as records too would duplicate what the template emits.  The Corefile's bind address
-        # derives from the IP as well, so CoreDNS has to come back on the new one.
-        await reload_coredns_for_domains(config, db)
-    else:
-        dns = DnsClient(db)
-        for domain in router_managed_domains(db):
-            await _point_at(dns, domain, detected)
-    return detected
-
-
-async def _point_at(dns: DnsClient, domain: str, ip: str) -> None:
-    """Point the names that follow the instance's address at ``ip``.
-
-    The apex and wildcard are what route the space and its apps; ``ns`` is the glue target when the
-    domain is delegated to us.  Which names those are is router policy, not something the DNS
-    client should know.
-    """
-    for fqdn in (domain, f"ns.{domain}", f"*.{domain}"):
-        await dns.set_records(fqdn, RecordType.A, [ip], ttl=_DYNAMIC_TTL_SECONDS)
-
-
-def start_dynamic_dns_task(
-    config: Config,
-    open_db: Callable[[], sqlite3.Connection],
-    interval_seconds: float = _DEFAULT_INTERVAL_SECONDS,
-) -> asyncio.Task[None]:
-    """Poll for address changes on the caller's event loop.
-
-    A fresh DB connection per tick, so a handle isn't held open across the sleep.  The caller must
-    keep the returned task alive — the loop holds only a weak reference.
-    """
-
-    async def _run() -> None:
-        logger.info(f"Dynamic DNS watcher started (every {interval_seconds:.0f}s)")
-        while True:
-            await asyncio.sleep(interval_seconds)
-            try:
-                with closing(open_db()) as db:
-                    await check_once(config, db)
-            except Exception:
-                # A transient provider or network error must not kill the watcher.
-                logger.exception("Dynamic DNS check failed; retrying at the next interval")
-
-    return asyncio.create_task(_run(), name="dynamic-dns")
